+++ database/migrations/2021_02_12_201246_create_all_loans_table.php
@@ -16,14 +16,14 @@ class CreateAllLoansTable extends Migration
         Schema::create('all_loans', function (Blueprint $table) {
             $table->id();
             $table->string('notes');
-            $table->integer('amount');
+            $table->double('amount');
             $table->longText('months_payable');
 
             $table->bigInteger('employee_id')->unsigned()->index();
             $table->foreign('employee_id')->references('id')
             ->on('employee')
             ->onDelete('cascade');
-            
+
             $table->bigInteger('loan_type_id')->unsigned()->index();
             $table->foreign('loan_type_id')->references('id')
             ->on('loan_type')
